Moved unit test files to separate directory; removed dead/obsolete code
