+++ BrowseRoyalCaribbeanPrice.py
@@ -1768,226 +1768,5 @@ def main(args: list[str] | None = None) -> None:
     log("Have a nice day!")
 
 
-##################################
-# Dead/Obsolete/Unused functions
-##################################
-'''
-appkey_mobile = 'cdCNc04srNq4rBvKofw1aC50dsdSaPuc' # royal
-appversion_mobile = '1.73.4'
-user_agent_mobile = 'royal/1.73.4 (com.rccl.royalcaribbean; build:2528; android 16) okhttp/4.12.0'
-
-def getShips():
-    headers = {
-        'appkey': appkey_mobile,
-        'accept': 'application/json',
-        'appversion': appversion_mobile,
-        'accept-language': 'en',
-        'user-agent': user_agent_mobile,
-    }
-
-    params = {
-        'sort': 'name',
-    }
-
-    try:
-        response = requests.get('https://api.rccl.com/en/all/mobile/v2/ships', params=params, headers=headers)
-    except Exception as e:
-        log(f"Can't contact cruise line servers; please try again later\n(program exception '{e}')")
-        exit(1)
-
-    shipNames = []
-    ships = response.json().get("payload").get("ships")
-    i = 0
-    for ship in ships:
-        shipCode = ship.get("shipCode")
-        name = ship.get("name")
-        shipNames.append({'code': shipCode, 'name': name})
-        if shipCode == "HM":
-            #Force Hero until it is added to the API
-            shipNames.append({'code': 'HE', 'name': 'Hero of the Seas'})
-
-    return shipNames
-
-def getSailings(shipCode):
-    headers = {
-        'appkey': appkey_mobile,
-        'accept': 'application/json',
-        'appversion': appversion_mobile,
-        'accept-language': 'en',
-        'user-agent': user_agent_mobile,
-    }
-
-    params = {
-        'resultSet': '300',
-    }
-
-    try:
-        response = requests.get(f'https://api.rccl.com/en/royal/mobile/v3/ships/{shipCode}/voyages', params=params, headers=headers)
-    except Exception as e:
-        log(f"Can't contact cruise line servers; please try again later\n(program exception '{e}')")
-        exit(1)
-
-    voyages = response.json().get("payload").get("voyages")
-    sailings = []
-    for voyage in voyages:
-        sailDate = voyage.get("sailDate")
-        duration = voyage.get("duration")
-        voyageCode = voyage.get("voyageCode")
-        voyageDescription = voyage.get("voyageDescription")
-        sailDateDisplay = datetime.strptime(sailDate, "%Y%m%d").strftime(DATE_DISPLAY_FORMAT)
-        sailings.append({'date': sailDate, 'displayDate': sailDateDisplay,'description': voyageDescription,'duration':duration,'voyageCode':voyageCode})
-
-    return sailings
-
-
-def getSailingDetails(shipCode,sailDate):
-    headers = {
-    'appkey': appkey_mobile,
-    'accept': 'application/json',
-    'user-agent': user_agent_mobile,
-    'appversion': appversion_mobile,
-    }
-
-    try:
-        response = requests.get(f'https://api.rccl.com/en/royal/mobile/v3/ships/voyages/{shipCode}{sailDate}/enriched', headers=headers)
-    except Exception as e:
-        log(f"Can't contact cruise line servers; please try again later\n(program exception '{e}')")
-        exit(1)
-
-    ports = {}
-
-    itinerary = response.json().get("payload").get("sailingInfo")[0].get("itinerary")
-    if itinerary is None:
-        return ports
-
-    portInfo = itinerary.get("portInfo")
-    for port in portInfo:
-        title = sanitize_string(port.get("title"))
-        arrivalDateTime = port.get("arrivalDateTime")
-        portType = port.get("portType","Unknown")
-        day = port.get("day","Unknown")
-
-        # Save port names for later
-        ports[int(day)] = title
-
-        # Sometimes fields are not filled, so we skip (like international date line changes)
-        if arrivalDateTime is None:
-            log(f"Day {day}: {title}")
-            continue
-
-        arrivalDateDisplay = datetime.strptime(arrivalDateTime[0:8], "%Y%m%d").strftime(DATE_DISPLAY_FORMAT)
-        departureDateTime = port.get("departureDateTime")
-        printString = f"Day {day} ({arrivalDateDisplay}): {title}"
-        if portType == "EMBARK":
-            printString += f" ↑ {departureDateTime[9:13]}"
-        elif portType == "DEBARK":
-            printString += f" ↓ {arrivalDateTime[9:13]}"
-        elif title != "Cruising":
-            printString += f" ↓ {arrivalDateTime[9:13]} ↑ {departureDateTime[9:13]}"
-        if portType == "TENDERED":
-            printString += f" (Tendered)"
-
-        portString = sanitize_string(printString)
-        log(portString)
-
-    if has_terminal_issues:
-        log("(^ means gangway up; v means gangway down)")
-    else:
-        log("(↑ means gangway up; ↓ means gangway down)")
-
-    return ports
-
-def getAllActivities(shipCode, sailDate):
-    headers = {
-        'appkey': appkey_mobile,
-        'accept': 'application/json',
-        'user-agent': user_agent_mobile,
-        'appversion': appversion_mobile,
-    }
-
-    params = {
-            'sailingID': shipCode + sailDate,
-            'limit': '200',
-            'offset': '0',
-            #'availableForSale': 'FALSE',
-        }
-
-    products = []
-    for offset in range(0,10000,200):
-        params['offset'] = offset
-
-        response = requests.get('https://api.rccl.com/en/royal/mobile/v3/products', params=params, headers=headers)
-        if response.json() is None or response.json().get("payload") is None:
-            return products
-
-        tempProducts = response.json().get("payload").get("products")
-        for product in tempProducts:
-            productType = product.get("productType").get("productType")
-
-            if productType != "NON_REVENUE_SCHEDULABLE":
-                continue
-
-            productTitle = product.get("productTitle")
-            location = product.get("productLocation").get("locationTitle","")
-
-            if location is None:
-                location = ""
-
-            offering = product.get("offering")
-            for offer in offering:
-                if offer is not None:
-                    offeringDate = offer.get("offeringDate")
-                    offeringTime = offer.get("offeringTime")
-                    day = days_between(sailDate,offeringDate)
-                    products.append({'productTitle':productTitle,'location':location,'offeringDate':offeringDate,'offeringTime':offeringTime,'day':day})
-
-    return products
-
-
-# This is not needed, as dress code is in list of activities
-def printThemeNights(shipCode,sailDate,duration):
-    headers = {
-        'appkey': appkey_mobile,
-        'accept': 'application/json',
-        'user-agent': user_agent_mobile,
-        'appversion': appversion_mobile,
-    }
-
-    response = requests.get(f'https://api.rccl.com/en/royal/mobile/v1/ships/{shipCode}/sailDate/{sailDate}/needToKnow', headers=headers)
-
-    payload = response.json().get("payload")
-
-    word_list = ["dress", "attire", "theme","casual"]
-
-    foundTheme = 0
-    for needToKnowCard in payload.get("needToKnowCards"):
-        includedDayCards = needToKnowCard.get("includedDayCards")
-        date = needToKnowCard.get("cardIdentifier")[:8]
-        for card in includedDayCards:
-            title = card.get("cardTitle")
-            titleLower = card.get("cardTitle").lower()
-            #if date == '20260402': # I use this to find missing themes
-            #    log(title)
-
-            if any(word.lower() in titleLower for word in word_list):
-                if "address" in titleLower:
-                    continue
-                foundTheme += 1
-
-                day = days_between(sailDate,date)
-                subtitle = re.split(r'[.!]+', card.get("cardSubtitle"))[0].replace("<p>", "").replace("&nbsp;","")
-                offeringDate = datetime.strptime(date, "%Y%m%d").strftime("%B %d, %Y")
-                log(f"{GREEN}{offeringDate} (Day {day}){RESET} {title}: {subtitle}")
-
-    if foundTheme == 0:
-        log("Themes Not Available")
-    elif foundTheme < duration:
-        log("Themes Not Fully Loaded")
-#    flush_print_buffer()
-'''
-##################################
-# End Dead/Obsolete/Unused functions
-##################################
-
 if __name__ == "__main__":
     main()
